requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework-simplejwt==5.4.0
django-environ==0.11.2
django-storages==1.14.4

# Channels & ASGI
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Database
psycopg2-binary==2.9.10

# Celery
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Redis
redis==5.2.1

# File handling
Pillow==11.0.0

# Search
django-postgres-extensions==0.9.3

# Utils
python-magic==0.4.27
gunicorn==23.0.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.24.0
factory-boy==3.3.1
